backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
# pydantic-email>=0.5.0
python-dotenv>=0.19.0
passlib>=1.7.4
bcrypt>=3.2.0
langchain>=0.0.267
openai>=0.28.1
psycopg2-binary>=2.9.1
# email-validator>=1.1.3
email-validator>=2.0.0
langchain_openai
